added create and update api for tools

+++ db/tools_dao.py
@@ -1,13 +1,20 @@
+from fastapi import HTTPException
 from db.models import Tools
+from exceptions.tool import ToolNotFoundError
 
 
 class ToolsDAO:
     def __init__(self, session):
         self.session = session
 
-    def get_all(self):
+    def get_all(self,limit : int=0,offset:int =0):
         """Get all tools."""
-        return self.session.query(Tools).all()
+        query = self.session.query(Tools)
+        if limit > 0:
+            query = query.limit(limit)
+        if offset > 0:
+            query = query.offset(offset)
+        return query.all()
     
     def get_by_id(self,id):
         """Get a tool by its ID."""
@@ -17,7 +24,6 @@ class ToolsDAO:
         """Create a new tool."""
         tool = Tools(**tool_data)
         self.session.add(tool)
-        self.session.commit()
         return tool
     
     def update(self, tool_id, updated_tool_data):
@@ -25,7 +31,7 @@ class ToolsDAO:
         tool = self.get_by_id(tool_id)
         for key, value in updated_tool_data.items():
             setattr(tool, key, value)
-        self.session.commit()
+        return tool
 
 
     def upsert(self, tool_data):
@@ -37,13 +43,13 @@ class ToolsDAO:
         else:
             new_tool = Tools(**tool_data)
             self.session.add(new_tool)
-            self.session.commit()
         return existing_tool or new_tool
 
 
     def delete(self, tool_id):
         """Delete a tool by its ID."""""
         tool = self.get_by_id(tool_id)
+        if not tool:
+            raise ToolNotFoundError(tool_id)
         self.session.delete(tool)
-        self.session.commit()
         return tool_id
